+++ eslint/babel-eslint-plugin-development/src/rules/no-deprecated-clone.js
@@ -1,7 +1,7 @@
-const getReferenceOrigin = require("../utils/get-reference-origin.cjs");
-const isFromBabelTypes = require("../utils/is-from-babel-types.cjs");
+import getReferenceOrigin from "../utils/get-reference-origin.js";
+import isFromBabelTypes from "../utils/is-from-babel-types.js";
 
-module.exports = {
+export default {
   meta: {
     schema: [],
     fixable: "code",
