+++ eslint/babel-eslint-plugin-development/src/utils/get-reference-origin.js
@@ -22,7 +22,7 @@ type ReferenceOrigin =
 // from.
 // It resolves imports, parameters of exported functions and property accesses.
 // See the ReferenceOrigin type for more information.
-module.exports = function getReferenceOrigin(
+export default function getReferenceOrigin(
   node,
   scope,
 ) /*: ?ReferenceOrigin */ {
@@ -69,7 +69,7 @@ module.exports = function getReferenceOrigin(
   }
 
   return null;
-};
+}
 
 function getVariableDefinition(name, scope) {
   let currentScope = scope;
